% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cohen}
\alias{cohen}
\title{Cohen's d}
\usage{
cohen(x, y)
}
\arguments{
\item{x, y}{numeric vectors containing the data of two groups.}
}
\description{
Computes cohens d from values of two independent groups
}
