% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Group values}
\usage{
group(x, n, type = "extreme")
}
\arguments{
\item{x}{numeric vector that is to be splitted-up in groups}

\item{n}{number of groups}
}
\value{
vector matching length of \code{x} binned into \code{n} groups.
}
\description{
Bin values in n groups.
}
\examples{
 # return mid points
 group(c(1,2,3,4,5),3,'mid')

 # return extreme points
 group(c(1,2,3,4,5),3,'extreme')

}
