% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replace_nc}
\alias{replace_nc}
\title{Replace numeric-character}
\usage{
replace_nc(elems, set, rep)
}
\arguments{
\item{elems}{numeric vector of to be replaced elements}

\item{set}{numeric vector containing the unique set of elems}

\item{rep}{character vector of same length as set containing the replacement values}
}
\value{
character vector of same length as \code{elems} containing the replacement values
}
\description{
Replace numeric values by character values.
}
\details{
Function iterates through \code{elems} and identifies a replacement in \code{rep} by
matching the elements to \code{set}
}
