% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{replace_nn}
\alias{replace_nn}
\title{Replace numeric-numeric}
\usage{
replace_nn(elems, set, rep)
}
\arguments{
\item{elems}{numeric vector of to be replaced elements}

\item{set}{numeric vector containing the unique set of elems}

\item{rep}{numeric vector of same length as set containing the replacement values}
}
\value{
numeric vector of same length as \code{elems} containing the replacement values
}
\description{
Replace numeric values by other numeric values.
}
\details{
Function iterates through \code{elems} and identifies a replacement in \code{rep} by
matching the elements to \code{set}
}
